moved util functions to separate file
