+++ packages/ui/Drawing/utils.js
@@ -6,7 +6,7 @@ const forAllNeighbors = (point, fn) => {
 	fn({x: point.x, y: point.y - 1});
 	fn({x: point.x + 1, y: point.y});
 	fn({x: point.x - 1, y: point.y});
-}
+};
 
 /*
  * Given two positions, returns true when they have a similar same color.
@@ -37,8 +37,54 @@ const relativePosition = (event, element) => {
 		y: Math.floor(event.clientY - rect.top)};
 };
 
+/*
+ * Executes the fill drawing on the canvas.
+ */
+const fillDrawing = (event, contextRef) => {
+	const canvas = contextRef.current.canvas;
+	const startPos = relativePosition(event, canvas);
+
+	const data = contextRef.current.getImageData(0, 0, canvas.width, canvas.height);
+	// An array with one place for each pixel in the image.
+	const alreadyFilled = new Array(data.width * data.height);
+
+	// This is the list of same-colored pixel coordinates as the starting point that were not handled yet.
+	const pixelList = [startPos];
+	while (pixelList.length) {
+		const pos = pixelList.pop();
+		const offset = pos.x + data.width * pos.y;
+		if (alreadyFilled[offset]) continue;
+
+		contextRef.current.fillRect(pos.x, pos.y, 1, 1);
+		alreadyFilled[offset] = true;
+
+		forAllNeighbors(pos, function (neighbor) {
+			if (neighbor.x >= 0 && neighbor.x < data.width &&
+				neighbor.y >= 0 && neighbor.y < data.height &&
+				isSimilarColor(data, startPos, neighbor)) {
+				pixelList.push(neighbor);
+			}
+		});
+	}
+};
+
+/*
+ * Executes the drawing on the canvas.
+ */
+const drawing = (beginPoint, controlPoint, endPoint, contextRef) => {
+	contextRef.current.beginPath();
+	contextRef.current.moveTo(beginPoint.x, beginPoint.y);
+	contextRef.current.quadraticCurveTo(
+		controlPoint.x,
+		controlPoint.y,
+		endPoint.x,
+		endPoint.y
+	);
+	contextRef.current.stroke();
+	contextRef.current.closePath();
+};
+
 export {
-	forAllNeighbors,
-	isSimilarColor,
-	relativePosition,
+	drawing,
+	fillDrawing
 };
